device-operations/src/Workers/requirements.txt
# Python Dependencies for SDXL Workers System
# Python 3.10.0 required
# For AMD RDNA RDNA2 RDNA3 use:
# AMD Software: Adrenalin Edition 23.40.27.06 for DirectML
# Core PyTorch with DirectML support
torch=2.4.1+cpu
torchvision=0.19.1+cpu
torch-directml=0.2.5.dev240914
open-clip-torch=2.32.0

# Diffusion models and transformers
diffusers=0.33.1
transformers=4.30.0
accelerate=0.20.0

# Model formats and safety
safetensors>=0.3.0
huggingface-hub>=0.16.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Scientific computing
numpy>=1.21.0
scipy>=1.9.0

# JSON schema validation
jsonschema>=4.17.0

# Async and utilities
aiofiles>=23.0.0
asyncio-throttle>=1.0.2

# Logging and monitoring
structlog>=23.0.0
psutil>=5.9.0

# Optional: Memory optimization
xformers>=0.0.20; platform_system != "Darwin"

# Optional: Performance monitoring
nvidia-ml-py>=12.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
